+++ e2e/src/whole_stack.rs
@@ -1,12 +1,12 @@
 use std::{sync::Arc, time::Duration};
 
-use storage::PostgresProcess;
+use storage::{Postgres, PostgresProcess};
 
 use crate::{
     committer::{Committer, CommitterProcess},
     eth_node::{ContractArgs, DeployedContract, EthNode, EthNodeProcess},
     fuel_node::{FuelNode, FuelNodeProcess},
-    kms::{Kms, KmsProcess},
+    kms::{Kms, KmsKey, KmsProcess},
 };
 
 #[allow(dead_code)]
@@ -22,65 +22,121 @@ pub struct WholeStack {
 
 impl WholeStack {
     pub async fn deploy_default(logs: bool, blob_support: bool) -> anyhow::Result<Self> {
-        let kms = Kms::default().with_show_logs(logs).start().await?;
-        let eth_node = EthNode::default().with_show_logs(logs).start().await?;
+        let kms = start_kms(logs).await?;
 
-        let amount = ethers::utils::parse_ether("400")?;
+        let eth_node = start_eth(logs).await?;
+        let (main_key, secondary_key) = create_and_fund_kms_keys(&kms, &eth_node).await?;
 
-        let main_wallet_key = kms.create_key(eth_node.chain_id()).await?;
-        eth_node.fund(main_wallet_key.address(), amount).await?;
+        let (contract_args, deployed_contract) = deploy_contract(&eth_node, &main_key).await?;
 
-        let contract_args = ContractArgs {
-            finalize_duration: Duration::from_secs(1),
-            blocks_per_interval: 10u32,
-            cooldown_between_commits: Duration::from_secs(1),
-        };
+        let fuel_node = start_fuel_node(logs).await?;
 
-        let deployed_contract = eth_node
-            .deploy_state_contract(main_wallet_key.clone(), contract_args)
-            .await?;
+        let (db_process, db) = start_db().await?;
 
-        let secondary_wallet_key = kms.create_key(eth_node.chain_id()).await?;
-        eth_node
-            .fund(secondary_wallet_key.address(), amount)
-            .await?;
+        let committer = start_committer(
+            logs,
+            blob_support,
+            db,
+            &eth_node,
+            &fuel_node,
+            &kms,
+            &deployed_contract,
+            &main_key,
+            &secondary_key,
+        )
+        .await?;
 
-        let fuel_node = FuelNode::default().with_show_logs(logs).start().await?;
-
-        let db_process = storage::PostgresProcess::shared().await?;
-        let random_db = db_process.create_random_db().await?;
-
-        let fuel_consensus_pub_key = fuel_node.consensus_pub_key();
-
-        let db_name = random_db.db_name();
-        let db_port = random_db.port();
-        let committer_builder = Committer::default()
-            .with_show_logs(logs)
-            .with_eth_rpc(eth_node.ws_url().clone())
-            .with_fuel_rpc(fuel_node.url().clone())
-            .with_db_port(db_port)
-            .with_db_name(db_name)
-            .with_state_contract_address(deployed_contract.address())
-            .with_fuel_block_producer_public_key(fuel_consensus_pub_key)
-            .with_main_key_id(main_wallet_key.id)
-            .with_aws_region(kms.region().clone());
-
-        let committer = if blob_support {
-            committer_builder.with_blob_key_id(secondary_wallet_key.id)
-        } else {
-            committer_builder
-        };
-
-        let committer = committer.start().await?;
-
-        Ok(Self {
+        Ok(WholeStack {
             eth_node,
             fuel_node,
             committer,
-            deployed_contract,
             db: db_process,
-            kms,
+            deployed_contract,
             contract_args,
+            kms,
         })
     }
+}
+
+async fn start_kms(logs: bool) -> anyhow::Result<KmsProcess> {
+    Kms::default().with_show_logs(logs).start().await
+}
+
+async fn start_eth(logs: bool) -> anyhow::Result<EthNodeProcess> {
+    EthNode::default().with_show_logs(logs).start().await
+}
+
+async fn create_and_fund_kms_keys(
+    kms: &KmsProcess,
+    eth_node: &EthNodeProcess,
+) -> anyhow::Result<(KmsKey, KmsKey)> {
+    let amount = ethers::utils::parse_ether("10")?;
+
+    let create_and_fund = || async {
+        let key = kms.create_key(eth_node.chain_id()).await?;
+        eth_node.fund(key.address(), amount).await?;
+        anyhow::Result::<_>::Ok(key)
+    };
+
+    Ok((create_and_fund().await?, create_and_fund().await?))
+}
+
+async fn deploy_contract(
+    eth_node: &EthNodeProcess,
+    main_wallet_key: &KmsKey,
+) -> anyhow::Result<(ContractArgs, DeployedContract)> {
+    let contract_args = ContractArgs {
+        finalize_duration: Duration::from_secs(1),
+        blocks_per_interval: 10u32,
+        cooldown_between_commits: Duration::from_secs(1),
+    };
+
+    let deployed_contract = eth_node
+        .deploy_state_contract(main_wallet_key.clone(), contract_args)
+        .await?;
+
+    Ok((contract_args, deployed_contract))
+}
+
+async fn start_fuel_node(logs: bool) -> anyhow::Result<FuelNodeProcess> {
+    FuelNode::default().with_show_logs(logs).start().await
+}
+
+async fn start_db() -> anyhow::Result<(Arc<PostgresProcess>, Postgres)> {
+    let db_process = storage::PostgresProcess::shared().await?;
+    let random_db = db_process.create_random_db().await?;
+
+    Ok((db_process, random_db))
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn start_committer(
+    logs: bool,
+    blob_support: bool,
+    random_db: Postgres,
+    eth_node: &EthNodeProcess,
+    fuel_node: &FuelNodeProcess,
+    kms: &KmsProcess,
+    deployed_contract: &DeployedContract,
+    main_key: &KmsKey,
+    secondary_key: &KmsKey,
+) -> anyhow::Result<CommitterProcess> {
+    let committer_builder = Committer::default()
+        .with_show_logs(logs)
+        .with_eth_rpc((eth_node).ws_url().clone())
+        .with_fuel_rpc(fuel_node.url().clone())
+        .with_db_port(random_db.port())
+        .with_db_name(random_db.db_name())
+        .with_state_contract_address(deployed_contract.address())
+        .with_fuel_block_producer_public_key(fuel_node.consensus_pub_key())
+        .with_main_key_id(main_key.id.clone())
+        .with_aws_region(kms.region().clone());
+
+    let committer = if blob_support {
+        committer_builder.with_blob_key_id(secondary_key.id.clone())
+    } else {
+        committer_builder
+    };
+
+    committer.start().await
 }
